requirements.txt
streamlit~=1.46.1
pypdf~=5.7.0
ollama~=0.5.1
protobuf~=5.29.5
plotly~=6.2.0
google-generativeai
python-dotenv